[KAN-60] Feat: 패키지 구조 수정
