+++ app/src/main/java/com/innerpeace/themoonha/adapter/BeforeAfterAdapter.kt
@@ -16,7 +16,7 @@ import com.google.android.exoplayer2.MediaItem
 import com.innerpeace.themoonha.R
 import com.innerpeace.themoonha.data.model.BeforeAfterContent
 import com.innerpeace.themoonha.databinding.FragmentBeforeAfterContentBinding
-import com.innerpeace.themoonha.ui.activity.beforeafter.view.BeforeAfterDetailFragment
+import com.innerpeace.themoonha.ui.fragment.beforeafter.BeforeAfterDetailFragment
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.launch
 import kotlinx.coroutines.withContext
