+++ app/src/main/java/com/innerpeace/themoonha/ui/fragment/craft/CraftFragment.kt
@@ -1,27 +1,26 @@
-package com.innerpeace.themoonha.ui.activity.lesson.view
+package com.innerpeace.themoonha.ui.fragment.craft
 
 import android.os.Bundle
-import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import com.innerpeace.themoonha.databinding.FragmentLessonDetailBinding
+import androidx.fragment.app.Fragment
+import com.innerpeace.themoonha.databinding.FragmentCraftBinding
 import com.innerpeace.themoonha.ui.activity.common.MainActivity
 
-class LessonDetailFragment : Fragment() {
-    private var _binding: FragmentLessonDetailBinding? = null
+class CraftFragment : Fragment() {
+    private var _binding: FragmentCraftBinding? = null
     private val binding get() = _binding!!
 
     override fun onCreateView(
-        inflater: LayoutInflater,
-        container: ViewGroup?,
+        inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
-        _binding = FragmentLessonDetailBinding.inflate(inflater, container, false)
+        _binding = FragmentCraftBinding.inflate(inflater, container, false)
         val view = binding.root
 
         // 툴바 제목 변경
-        (activity as? MainActivity)?.setToolbarTitle("강좌 상세")
+        (activity as? MainActivity)?.setToolbarTitle("문화공방")
 
         return view
     }
