+++ app/src/main/java/com/innerpeace/themoonha/ui/fragment/lesson/LessonDetailFragment.kt
@@ -1,26 +1,27 @@
-package com.innerpeace.themoonha.ui.activity.live.view
+package com.innerpeace.themoonha.ui.fragment.lesson
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import com.innerpeace.themoonha.databinding.FragmentLiveBinding
+import com.innerpeace.themoonha.databinding.FragmentLessonDetailBinding
 import com.innerpeace.themoonha.ui.activity.common.MainActivity
 
-class LiveFragment : Fragment() {
-    private var _binding: FragmentLiveBinding? = null
+class LessonDetailFragment : Fragment() {
+    private var _binding: FragmentLessonDetailBinding? = null
     private val binding get() = _binding!!
 
     override fun onCreateView(
-        inflater: LayoutInflater, container: ViewGroup?,
+        inflater: LayoutInflater,
+        container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
-        _binding = FragmentLiveBinding.inflate(inflater, container, false)
+        _binding = FragmentLessonDetailBinding.inflate(inflater, container, false)
         val view = binding.root
 
         // 툴바 제목 변경
-        (activity as? MainActivity)?.setToolbarTitle("실시간 강좌")
+        (activity as? MainActivity)?.setToolbarTitle("강좌 상세")
 
         return view
     }
