+++ app/src/main/java/com/innerpeace/themoonha/ui/fragment/lesson/LessonFragment.kt
@@ -1,10 +1,10 @@
-package com.innerpeace.themoonha.ui.activity.lesson.view
+package com.innerpeace.themoonha.ui.fragment.lesson
 
 import android.os.Bundle
-import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import androidx.fragment.app.Fragment
 import androidx.navigation.fragment.findNavController
 import com.innerpeace.themoonha.R
 import com.innerpeace.themoonha.databinding.FragmentLessonBinding
