+++ app/src/main/java/com/innerpeace/themoonha/ui/fragment/live/LiveFragment.kt
@@ -1,27 +1,26 @@
-package com.innerpeace.themoonha.ui.activity.lesson.view
+package com.innerpeace.themoonha.ui.fragment.live
 
 import android.os.Bundle
-import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import com.innerpeace.themoonha.databinding.FragmentCartBinding
+import androidx.fragment.app.Fragment
+import com.innerpeace.themoonha.databinding.FragmentLiveBinding
 import com.innerpeace.themoonha.ui.activity.common.MainActivity
 
-
-class CartFragment : Fragment() {
-    private var _binding: FragmentCartBinding? = null
+class LiveFragment : Fragment() {
+    private var _binding: FragmentLiveBinding? = null
     private val binding get() = _binding!!
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
-        _binding = FragmentCartBinding.inflate(inflater, container, false)
+        _binding = FragmentLiveBinding.inflate(inflater, container, false)
         val view = binding.root
 
         // 툴바 제목 변경
-        (activity as? MainActivity)?.setToolbarTitle("나의 문화센터")
+        (activity as? MainActivity)?.setToolbarTitle("실시간 강좌")
 
         return view
     }
