+++ app/src/main/java/com/innerpeace/themoonha/ui/fragment/lounge/LoungeFragment.kt
@@ -1,15 +1,15 @@
-package com.innerpeace.themoonha.ui.activity.schedule.view
+package com.innerpeace.themoonha.ui.fragment.lounge
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import com.innerpeace.themoonha.databinding.FragmentScheduleBinding
+import com.innerpeace.themoonha.databinding.FragmentLoungeBinding
 import com.innerpeace.themoonha.ui.activity.common.MainActivity
 
 /**
- * 스케줄 프래그먼트
+ * 라운지 프래그먼트
  * @author 조희정
  * @since 2024.08.23
  * @version 1.0
@@ -20,8 +20,9 @@ import com.innerpeace.themoonha.ui.activity.common.MainActivity
  * 2024.08.24  	조희정       최초 생성
  * </pre>
  */
-class ScheduleFragment : Fragment() {
-    private var _binding: FragmentScheduleBinding? = null
+class LoungeFragment : Fragment() {
+
+    private var _binding: FragmentLoungeBinding? = null
     private val binding get() = _binding!!
 
     override fun onCreateView(
@@ -29,11 +30,11 @@ class ScheduleFragment : Fragment() {
         container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
-        _binding = FragmentScheduleBinding.inflate(inflater, container, false)
+        _binding = FragmentLoungeBinding.inflate(inflater, container, false)
         val view = binding.root
 
         // 툴바 제목 변경
-        (activity as? MainActivity)?.setToolbarTitle("스케줄")
+        (activity as? MainActivity)?.setToolbarTitle("라운지")
 
         return view
     }
@@ -42,5 +43,4 @@ class ScheduleFragment : Fragment() {
         super.onDestroyView()
         _binding = null
     }
-
 }
