+++ app/src/main/java/com/innerpeace/themoonha/ui/fragment/lounge/LoungeHomeFragment.kt
@@ -1,15 +1,15 @@
-package com.innerpeace.themoonha.ui.activity.lounge.view
+package com.innerpeace.themoonha.ui.fragment.lounge
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import com.innerpeace.themoonha.databinding.FragmentLoungeBinding
+import com.innerpeace.themoonha.databinding.FragmentLoungeHomeBinding
 import com.innerpeace.themoonha.ui.activity.common.MainActivity
 
 /**
- * 라운지 프래그먼트
+ * 라운지 개별 홈 프래그먼트
  * @author 조희정
  * @since 2024.08.23
  * @version 1.0
@@ -20,9 +20,9 @@ import com.innerpeace.themoonha.ui.activity.common.MainActivity
  * 2024.08.24  	조희정       최초 생성
  * </pre>
  */
-class LoungeFragment : Fragment() {
+class LoungeHomeFragment : Fragment() {
 
-    private var _binding: FragmentLoungeBinding? = null
+    private var _binding: FragmentLoungeHomeBinding? = null
     private val binding get() = _binding!!
 
     override fun onCreateView(
@@ -30,11 +30,11 @@ class LoungeFragment : Fragment() {
         container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
-        _binding = FragmentLoungeBinding.inflate(inflater, container, false)
+        _binding = FragmentLoungeHomeBinding.inflate(inflater, container, false)
         val view = binding.root
 
         // 툴바 제목 변경
-        (activity as? MainActivity)?.setToolbarTitle("라운지")
+        (activity as? MainActivity)?.setToolbarTitle("라운지 개별 이름")
 
         return view
     }
@@ -43,4 +43,5 @@ class LoungeFragment : Fragment() {
         super.onDestroyView()
         _binding = null
     }
+
 }
