+++ app/src/main/java/com/innerpeace/themoonha/ui/fragment/lounge/LoungePostFragment.kt
@@ -1,15 +1,15 @@
-package com.innerpeace.themoonha.ui.activity.lounge.view
+package com.innerpeace.themoonha.ui.fragment.lounge
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import com.innerpeace.themoonha.databinding.FragmentLoungePostWriteBinding
+import com.innerpeace.themoonha.databinding.FragmentLoungePostBinding
 import com.innerpeace.themoonha.ui.activity.common.MainActivity
 
 /**
- * 라운지 게시글 작성 프래그먼트
+ * 라운지 게시글 상세 프래그먼트
  * @author 조희정
  * @since 2024.08.23
  * @version 1.0
@@ -20,19 +20,20 @@ import com.innerpeace.themoonha.ui.activity.common.MainActivity
  * 2024.08.24  	조희정       최초 생성
  * </pre>
  */
-class LoungePostWriteFragment : Fragment() {
-    private var _binding: FragmentLoungePostWriteBinding? = null
+class LoungePostFragment : Fragment() {
+    private var _binding: FragmentLoungePostBinding? = null
     private val binding get() = _binding!!
 
     override fun onCreateView(
-        inflater: LayoutInflater, container: ViewGroup?,
+        inflater: LayoutInflater,
+        container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
-        _binding = FragmentLoungePostWriteBinding.inflate(inflater, container, false)
+        _binding = FragmentLoungePostBinding.inflate(inflater, container, false)
         val view = binding.root
 
         // 툴바 제목 변경
-        (activity as? MainActivity)?.setToolbarTitle("글 작성")
+        (activity as? MainActivity)?.setToolbarTitle("라운지 개별 이름")
 
         return view
     }
