+++ app/src/main/java/com/innerpeace/themoonha/ui/fragment/lounge/LoungePostWriteFragment.kt
@@ -1,15 +1,15 @@
-package com.innerpeace.themoonha.ui.activity.lounge.view
+package com.innerpeace.themoonha.ui.fragment.lounge
 
 import android.os.Bundle
-import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import com.innerpeace.themoonha.databinding.FragmentLoungeHomeBinding
+import androidx.fragment.app.Fragment
+import com.innerpeace.themoonha.databinding.FragmentLoungePostWriteBinding
 import com.innerpeace.themoonha.ui.activity.common.MainActivity
 
 /**
- * 라운지 개별 홈 프래그먼트
+ * 라운지 게시글 작성 프래그먼트
  * @author 조희정
  * @since 2024.08.23
  * @version 1.0
@@ -20,21 +20,19 @@ import com.innerpeace.themoonha.ui.activity.common.MainActivity
  * 2024.08.24  	조희정       최초 생성
  * </pre>
  */
-class LoungeHomeFragment : Fragment() {
-
-    private var _binding: FragmentLoungeHomeBinding? = null
+class LoungePostWriteFragment : Fragment() {
+    private var _binding: FragmentLoungePostWriteBinding? = null
     private val binding get() = _binding!!
 
     override fun onCreateView(
-        inflater: LayoutInflater,
-        container: ViewGroup?,
+        inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
-        _binding = FragmentLoungeHomeBinding.inflate(inflater, container, false)
+        _binding = FragmentLoungePostWriteBinding.inflate(inflater, container, false)
         val view = binding.root
 
         // 툴바 제목 변경
-        (activity as? MainActivity)?.setToolbarTitle("라운지 개별 이름")
+        (activity as? MainActivity)?.setToolbarTitle("글 작성")
 
         return view
     }
