+++ app/src/main/java/com/innerpeace/themoonha/ui/fragment/schedule/ScheduleFragment.kt
@@ -1,15 +1,15 @@
-package com.innerpeace.themoonha.ui.activity.lounge.view
+package com.innerpeace.themoonha.ui.fragment.schedule
 
 import android.os.Bundle
-import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import com.innerpeace.themoonha.databinding.FragmentLoungePostBinding
+import androidx.fragment.app.Fragment
+import com.innerpeace.themoonha.databinding.FragmentScheduleBinding
 import com.innerpeace.themoonha.ui.activity.common.MainActivity
 
 /**
- * 라운지 게시글 상세 프래그먼트
+ * 스케줄 프래그먼트
  * @author 조희정
  * @since 2024.08.23
  * @version 1.0
@@ -20,8 +20,8 @@ import com.innerpeace.themoonha.ui.activity.common.MainActivity
  * 2024.08.24  	조희정       최초 생성
  * </pre>
  */
-class LoungePostFragment : Fragment() {
-    private var _binding: FragmentLoungePostBinding? = null
+class ScheduleFragment : Fragment() {
+    private var _binding: FragmentScheduleBinding? = null
     private val binding get() = _binding!!
 
     override fun onCreateView(
@@ -29,11 +29,11 @@ class LoungePostFragment : Fragment() {
         container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
-        _binding = FragmentLoungePostBinding.inflate(inflater, container, false)
+        _binding = FragmentScheduleBinding.inflate(inflater, container, false)
         val view = binding.root
 
         // 툴바 제목 변경
-        (activity as? MainActivity)?.setToolbarTitle("라운지 개별 이름")
+        (activity as? MainActivity)?.setToolbarTitle("스케줄")
 
         return view
     }
